[spu_app]send to and receive from ciu98m25 spu via uart transparent transfer. [known issue]max send data to spu is 255 bytes,max receive data from spu is 253 bytes. [info]code=3919 const=448 data=942

+++ User/Uart.h
@@ -12,7 +12,7 @@ extern void Uart1PutUint8AsDecimal(uint8 dat);
 extern OS_SEM_t SemUart1Tx;
 extern OS_SEM_t SemUart1RxOver;//串口1一帧数据已接收完成(接收超时)
 extern uint8 Uart1RxBuff[];
-extern uint8 Uart1RxPtr;//串口发送和接收共用的缓冲指针
+extern uint16 Uart1RxPtr;//串口发送和接收共用的缓冲指针
 //*************************************************************
 #define SBUF                     UART1->DR//51形式的宏定义
 #define WAIT_UART_TX_IDLE()      while(!GBF(UART1->SR,UART_SR_TXE))//等待串口可以发送
